Add reorder loop, exercise functionalities

+++ core/data/src/main/java/com/wp/core/data/repository/WorkoutRepo.kt
@@ -5,6 +5,6 @@ import com.comsuon.wp.model.WorkoutModel
 interface WorkoutRepo {
     suspend fun getWorkoutDataList(): List<WorkoutModel>
     suspend fun saveWorkoutData(data: WorkoutModel)
-    suspend fun getWorkout(id: Int): WorkoutModel
-    suspend fun deleteWorkout(id: Int)
+    suspend fun getWorkout(id: Long): WorkoutModel
+    suspend fun deleteWorkout(id: Long)
 }
