+++ core/model/src/main/java/com/comsuon/wp/model/WorkoutModel.kt
@@ -1,7 +1,7 @@
 package com.comsuon.wp.model
 
 data class WorkoutModel(
-    var index: Int = 0,
+    var index: Long = 0,
     var workoutName: String = "",
     var loopList: List<LoopModel> = listOf(),
     var isFavourite: Boolean = false
